Moved Result utility methods from ALISA into ResultUtil and removed them within ALISA.
