+++ core/org.osate.results/src/org/osate/result/util/ResultUtil.java
@@ -1,5 +1,7 @@
 package org.osate.result.util;
 
+import java.util.Collection;
+
 import org.eclipse.emf.common.util.EList;
 import org.eclipse.emf.ecore.EObject;
 import org.osate.result.AnalysisResult;
@@ -47,6 +49,13 @@ public class ResultUtil {
 		return issue;
 	}
 
+	public static Result createResult(String name, EObject target) {
+		Result result = ResultFactory.eINSTANCE.createResult();
+		result.setSourceReference(target);
+		result.setInfo(name);
+		return result;
+	}
+
 	public static AnalysisResult createAnalysisResult(String name, EObject target) {
 		AnalysisResult result = ResultFactory.eINSTANCE.createAnalysisResult();
 		result.setSourceReference(target);
@@ -91,6 +100,8 @@ public class ResultUtil {
 		res.setValue(value);
 		contributor.getValues().add(res);
 	}
+
+	// get values from the Result object
 
 	public static String getString(Result result, int idx) {
 		EList<Value> vals = result.getValues();
@@ -143,4 +154,95 @@ public class ResultUtil {
 		}
 		return null;
 	}
+
+	/**
+	 * true if there are any diagnostics with FAILURE type or with ERROR type
+	 * @param res
+	 * @return
+	 */
+	public static boolean hasFailures(Result res) {
+		if (hasFailures(res.getDiagnostics())) {
+			return true;
+		}
+		for (Result subres : res.getSubResults()) {
+			if (hasFailures(subres)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasFailures(AnalysisResult res) {
+		for (Result r : res.getResults()) {
+			if (hasFailures(r)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasFailures(Diagnostic res) {
+		if (res.getType() == DiagnosticType.FAILURE) {
+			return true;
+		}
+		// this code supports nested Diagnostic objects
+		for (Diagnostic r : res.getIssues()) {
+			if (hasFailures(r)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasFailures(Collection<Diagnostic> res) {
+		for (Diagnostic r : res) {
+			if (hasFailures(r)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasErrors(Collection<Diagnostic> res) {
+		for (Diagnostic r : res) {
+			if (hasErrors(r)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasErrors(Diagnostic res) {
+		if (res.getType() == DiagnosticType.ERROR) {
+			return true;
+		}
+		for (Diagnostic r : res.getIssues()) {
+			if (hasFailures(r)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasErrors(Result res) {
+		if (hasErrors(res.getDiagnostics())) {
+			return true;
+		}
+		for (Result subres : res.getSubResults()) {
+			if (hasErrors(subres)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
+	public static boolean hasErrors(AnalysisResult res) {
+		for (Result r : res.getResults()) {
+			if (hasErrors(r)) {
+				return true;
+			}
+		}
+		return false;
+	}
+
 }
